refactor: OfficalQuestion 엔티티 경로 변경 및 속성 수정
